+++ cli/cli/helpers/host_machine_directories/host_machine_directories.go
@@ -17,7 +17,7 @@ const (
 
 	metricsUserIDFilename = "metrics-user-id"
 
-	userConsentToSendMetricsElection = "user-consent-to-send-metrics-election"
+	userSendMetricsElection = "user-send-metrics-election"
 
 	// ------------ Names of dirs inside Kurtosis directory --------------
 	engineDataDirname = "engine-data"
@@ -52,11 +52,11 @@ func GetMetricsUserIdFilepath() (string, error) {
 	return filepath, nil
 }
 
-func GetUserConsentToSendMetricsElectionFilepath() (string, error) {
-	xdgRelFilepath := getRelativeFilepathForXDG(userConsentToSendMetricsElection)
-	filepath, err := xdg.ConfigFile(xdgRelFilepath)
+func GetUserSendMetricsElectionFilepath() (string, error) {
+	xdgRelFilepath := getRelativeFilepathForXDG(userSendMetricsElection)
+	filepath, err := xdg.DataFile(xdgRelFilepath)
 	if err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred getting the user consent to send metrics election filepath from relative path '%v'", xdgRelFilepath)
+		return "", stacktrace.Propagate(err, "An error occurred getting the user-send-metrics-election filepath from relative path '%v'", xdgRelFilepath)
 	}
 	return filepath, nil
 }