+++ cli/internal_testsuites/golang/testsuite/network_soft_partition_test/network_soft_partition_test.go
@@ -94,7 +94,7 @@ func TestNetworkSoftPartitions(t *testing.T) {
 		"--report",
 		"--json",
 		"--report-cycles",
-		"1", // We set report cycles to 1 to generate the report faster because default is 10
+		"2", // We set report cycles to 2 to generate the report faster because default is 10
 		"--no-dns", //No domain name resolution, also to improve velocity
 	}
 
@@ -172,7 +172,7 @@ func TestNetworkSoftPartitions(t *testing.T) {
 	mtrReportAfterUnblockedPartition := new(MtrReport)
 	err = json.Unmarshal([]byte(jsonStr), mtrReportAfterUnblockedPartition)
 	require.NoError(t, err, "An error occurred unmarshalling json string '%v' to mtr report struct ", jsonStr)
-	require.Greaterf(t, len(mtrReportAfterUnblockedPartition.Report.Hubs), zeroElementsInMtrHubField, "There isn't any element in the report hub field")
+	require.Greaterf(t, len(mtrReportAfterUnblockedPartition.Report.Hubs), zeroElementsInMtrHubField, "There aren't any element in the report hub field")
 	require.Equal(t, zeroPacketLoss, mtrReportAfterUnblockedPartition.Report.Hubs[0].Loss)
 	logrus.Info("Report complete successfully, there was no packet loss between services during the test")
 }